decodebin: Manually sync element states with parent after unsetting locked_state
When the pads of the new elements are connected in decodebin, the new element
can remain in PAUSED state while decodebin is already in PLAYING. Decodebin
isn't responsible anymore to set their children to PLAYING, so nobody will do
it.

This patch detects that situation and manually synchronizes the state of the
new element to the same of its parent (decodebin).

This change helps to solve some stalls in HLS videos when combined with
certain hardware decoders.